minor bug fixes, api changes, double support

+++ binpack.h
@@ -1,3 +1,4 @@
+/* SAS modified this file. */
 /* (C) Copyright 2003 Jens Lysgaard. All rights reserved. */
 /* OSI Certified Open Source Software */
 /* This software is licensed under the Common Public License Version 1.0 */
@@ -5,8 +6,8 @@
 #ifndef _H_BINPACK
 #define _H_BINPACK
 
-void BP_ExactBinPacking(int CAP,
-                        int *ItemSize,
+void BP_ExactBinPacking(double CAP,
+                        double *ItemSize,
                         int n,
                         int *LB,
                         int *UB,
