+++ twomatch.cpp
@@ -1,3 +1,4 @@
+/* SAS modified this file. */
 /* (C) Copyright 2003 Jens Lysgaard. All rights reserved. */
 /* OSI Certified Open Source Software */
 /* This software is licensed under the Common Public License Version 1.0 */
@@ -256,7 +257,7 @@ void TWOMATCH_ExactTwoMatchings(ReachPtr SupportPtr,
     Head = NewNodeHead[i];
 
     XVal = XMatrix[Tail][Head];
-    ArcCap = XVal * Scale;
+    ArcCap = (int)(XVal * Scale);
     if (ArcCap < 0) ArcCap = 0;
     if (ArcCap > MaxArcCap) ArcCap = MaxArcCap;
 
@@ -286,7 +287,7 @@ void TWOMATCH_ExactTwoMatchings(ReachPtr SupportPtr,
         are not complemented) */
 
     XVal = XMatrix[NoOfCustomers+1][k];
-    ArcCap = XVal * Scale;
+    ArcCap = (int)(XVal * Scale);
 
     if (ArcCap > 0)
     {
